Create an information table to record total row counts of each table and avoid to import the same data repeatedly.

+++ red/com/xl/parsers/referenceparsers/DNAVCFParser.java
@@ -15,6 +15,7 @@ package com.xl.parsers.referenceparsers;
 
 import com.xl.database.DatabaseManager;
 import com.xl.interfaces.ProgressListener;
+import com.xl.utils.EmptyChecker;
 import com.xl.utils.Indexer;
 
 import java.io.BufferedReader;
@@ -54,7 +55,7 @@ public class DNAVCFParser extends AbstractParser {
     private DatabaseManager databaseManager = DatabaseManager.getInstance();
 
     public DNAVCFParser(String dataPath) {
-        super(dataPath);
+        super(dataPath, DatabaseManager.DNA_VCF_RESULT_TABLE_NAME);
     }
 
     @Override
@@ -70,7 +71,6 @@ public class DNAVCFParser extends AbstractParser {
             String line;
             String[] columnStrings = new String[0];
             StringBuilder tableBuilders = new StringBuilder();
-            databaseManager.setAutoCommit(false);
             int lineCount = 0;
             boolean hasEstablishTable = false;
             while ((line = bufferedReader.readLine()) != null) {
@@ -82,10 +82,9 @@ public class DNAVCFParser extends AbstractParser {
                     sampleNames = Arrays.copyOfRange(columnStrings, formatColumnIndex + 1, columnLength);
                     tableBuilders.append(columnStrings[0]).append(" varchar(30),").append(columnStrings[1])
                         .append(" int,").append(columnStrings[2]).append(" varchar(30),").append(columnStrings[3])
-                        .append(" varchar(5),").append(columnStrings[4]).append(" varchar(5),")
-                        .append(columnStrings[5]).append(" float(10,2),").append(columnStrings[6]).append(" text,")
-                        .append(columnStrings[7]).append(" text,");
-                    // return;
+                        .append(" varchar(5),").append(columnStrings[4]).append(" varchar(5),").append(columnStrings[5])
+                        .append(" float(10,2),").append(columnStrings[6]).append(" text,").append(columnStrings[7])
+                        .append(" text,");
                     continue;
                 }
                 if (sampleNames == null) {
@@ -121,11 +120,17 @@ public class DNAVCFParser extends AbstractParser {
                         tableNames = new String[sampleNames.length];
                         for (int j = 0, len = sampleNames.length; j < len; j++) {
                             tableNames[j] = sampleNames[j] + "_" + DatabaseManager.DNA_VCF_RESULT_TABLE_NAME;
-                            databaseManager.deleteTable(tableNames[j]);
-                            databaseManager.executeSQL("create table " + tableNames[j] + "(" + tableBuilders + ")");
                         }
-                        databaseManager.commit();
-                        hasEstablishTable = true;
+                        if (isDataValid(tableNames)) {
+                            return;
+                        } else {
+                            for (int j = 0, len = sampleNames.length; j < len; j++) {
+                                databaseManager.deleteTable(tableNames[j]);
+                                databaseManager.executeSQL("create table " + tableNames[j] + "(" + tableBuilders + ")");
+                            }
+                            databaseManager.setAutoCommit(false);
+                            hasEstablishTable = true;
+                        }
                     }
 
                     // INSERT INTO table_name (col1, col2,...) VALUES (value1, value2,....)
@@ -197,6 +202,23 @@ public class DNAVCFParser extends AbstractParser {
                 }
             }
         }
+    }
+
+    @Override
+    protected void recordInformation() {
+        if (!EmptyChecker.isEmptyArray(tableNames)) {
+            for (String tableName : tableNames) {
+                databaseManager.insertOrUpdateInfo(tableName);
+            }
+        }
+    }
+
+    private boolean isDataValid(String[] tableNames) {
+        boolean valid = true;
+        for (String tableName : tableNames) {
+            valid &= databaseManager.isTableExistAndValid(tableName);
+        }
+        return valid;
     }
 
     public String[] getSampleNames() {
